Database script wooorks!!  Still not sure if it's ok for it to be all in the same script.
